#4 Moved static method from Samples class as the source code parser tripped over it
