+++ README.md
@@ -566,34 +566,6 @@ public Table md() {
 			.append(row(cell("Cell 2.1"), cell(italic("Cell 2.2")))) //
 			.build();
 }
-public static List<Method> sampleMethods() {
-	return Arrays.stream(Samples.class.getDeclaredMethods())
-			// Only methods that are marked as Sample methods
-			.filter(m -> m.getAnnotation(Sample.class) != null)
-			/**
-			 * By definition "sample" methods have no parameters and return
-			 * a MarkdownSerilizable. Skip all methods that do not meet
-			 * those criteria.
-			 */
-			.filter(m -> {
-				// Only no-arg methods
-				if (m.getParameterCount() > 0) {
-					return false;
-				}
-				// Only BlockElements or SpanElements
-				if (BlockElement.class.isAssignableFrom(m.getReturnType())) {
-					return true;
-				}
-				if (SpanElement.class.isAssignableFrom(m.getReturnType())) {
-					return true;
-				}
-				return false;
-			})
-			// Sort by order defined by annotation
-			.sorted(Comparator.comparingInt(m -> m.getAnnotation(Sample.class).order()))
-			// Capture the sorted list of methods.
-			.collect(Collectors.toList());
-}
 ```
 
 ### Markdown
